requirements.txt
numpy
pandas
python-dotenv
psycopg2-binary
scikit-learn
seaborn
catboost
xgboost
mlflow
#-e .